requirements.txt
PyYAML>=6.0
phue==1.1
requests>=2.31.0
httpx>=0.24.0
Flask>=2.3.0
playwright>=1.49.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.21.0
